MYR-144 fix(store): split writer.go (line cap), context scoping, (0,0) sentinel
Three Medium-severity Claude Review findings on PR #253:

1. writer.go was 333 lines (CLAUDE.md 300-line file cap). Moved
   handleTelemetry, flushLoop, and flush into writer_flush.go;
   writer.go drops to 255 lines.

2. runLocationGeocode and geocodeParkedLocation shared a single 5s
   context between the Mapbox call and the follow-up DB write — a
   4.5s geocode would leave ~500ms for UpdateTelemetry. Split into
   two scoped contexts using a new locationDBWriteTimeout sibling.

3. scheduleLocationGeocode didn't filter the (0,0) GPS sentinel,
   asymmetric with geocodeParkedLocation and the drive-start/end
   paths. Added the guard plus a covering test.
