+++ internal/store/writer.go
@@ -253,81 +253,3 @@ func (w *Writer) Stop() error {
 func (w *Writer) Done() <-chan struct{} {
 	return w.done
 }
-
-// handleTelemetry is the event handler for VehicleTelemetryEvent. It
-// extracts fields, maps them to a VehicleUpdate, and coalesces into
-// the pending map. If the batch size is reached, it triggers a flush.
-func (w *Writer) handleTelemetry(event events.Event) {
-	telEvt, ok := event.Payload.(events.VehicleTelemetryEvent)
-	if !ok {
-		w.logger.Error("unexpected payload type for telemetry event",
-			slog.String("event_id", event.ID),
-		)
-		return
-	}
-
-	update := mapTelemetryToUpdate(telEvt.Fields)
-	if update == nil {
-		return
-	}
-	update.LastUpdated = telEvt.CreatedAt
-
-	shouldFlush := w.coalesce(telEvt.VIN, update)
-	if shouldFlush {
-		ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
-		defer cancel()
-		w.flush(ctx)
-	}
-}
-
-// flushLoop runs a ticker that calls flush at each interval until the
-// context is cancelled.
-func (w *Writer) flushLoop(ctx context.Context) {
-	ticker := time.NewTicker(w.cfg.FlushInterval)
-	defer ticker.Stop()
-
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		case <-ticker.C:
-			w.flush(ctx)
-		}
-	}
-}
-
-// flush drains the pending map and writes each VIN's coalesced update
-// to the database. Errors are logged but do not halt the writer.
-func (w *Writer) flush(ctx context.Context) {
-	w.pendingMu.Lock()
-	if len(w.pending) == 0 {
-		w.pendingMu.Unlock()
-		return
-	}
-	batch := w.pending
-	w.pending = make(map[string]*VehicleUpdate)
-	w.count = 0
-	w.pendingMu.Unlock()
-
-	w.logger.Debug("flushing telemetry batch",
-		slog.Int("vehicles", len(batch)),
-	)
-
-	for vin, update := range batch {
-		w.applyDestinationAddress(ctx, vin, update)
-		if err := w.vehicles.UpdateTelemetry(ctx, vin, *update); err != nil {
-			w.logger.Warn("failed to write telemetry update",
-				slog.String("vin", redactVIN(vin)),
-				slog.String("error", err.Error()),
-			)
-			continue
-		}
-		// MYR-144: schedule an async reverse-geocode of the current
-		// location so Vehicle.locationName / Vehicle.locationAddress
-		// stay fresh without blocking the flush loop. The scheduler
-		// applies a per-VIN time + distance debounce, so a stable
-		// parked vehicle does not burn the Mapbox quota.
-		w.scheduleLocationGeocode(vin, update)
-	}
-}
-
